requirements.txt
python-dotenv
zhipuai